server/requirements.txt
fastapi
uvicorn
python-dotenv
pandas
hydra-core
omegaconf
rich
python-multipart
openai
seaborn